// src/shared/generated/AppKitRuntimeProjectionEffectRecord.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { AppKitRuntimeWindowProjectionRecord } from "./AppKitRuntimeWindowProjectionRecord";

export type AppKitRuntimeProjectionEffectRecord = { contentFocusTabId?: string, eventId: string, windows: Array<AppKitRuntimeWindowProjectionRecord>, };